Add files via upload
First action in coding a new blockchain for a only cryptocurrecy and bitcoin buy and sell providing new/used physical/digital items growing a bigger value to more availability of online shopping with your hard earned crypto or bitcoin assets the blockchain has a Sell Block A purchase Block A delivery Block and a final confirmation Block Human Hash function and cpu hash function to secure and approve item has been accepted by the buyer is reward in a 50/30/20 scale as to Miner/Charity/Business reward perspective! a small transaction fee will be applied by the Block Dispensary Blockchain to secure transactions and 50 percent of the transaction fee goin to Miners 30 percent going to global and local charities and 20 percent held by the Block Dispensary business for development and long term operate security providing a web buy and sell for all items with only bitcoin and crypto currency!