Create SellPage.js
A first commit of the blockchain dispensary buy/sell SELL PAGE. The sell page will have top page listings like all other pages in the header and it the body 1-3 images will appear in the body advertising items for sell on the blockchain! There is a list item for sell link to another page option to use social profile to list item for sell and submit the item to the first sector of approval of item human verified hash earning a reward from a confirmation on the block dispensary website! there is a click buy item button just incase you see an item to buy while going to list an item for sell! im javascript skills lack any help or suggestions would be appreciated i am just using the late night early morning to write the basic HTML route to outline the project!