Create BrowseItems.js
the Browse Items for sale on the Block dispensary Blockchain page with 3 images by 2 columns  could be more columns and rows changing items every 5 - 10 seconds revolving on a scale displaying a new image out of the 6 displayed every 5 seconds giving an option to click on the photo and be redirected to the item purchase page! the revolving to the right as far until out of page view and then on the left displaying new image! a  link to list item on the browse page to be redirected to list to sell page! A brief mention of social networking on the website and that us as block dispensary do not sell personal data to governments or corporations mention of evolving digital currency demand opening up more options to purchase items with online digital currencies!