Create ListItem.js
Here is the list item page consisting of credentials of listing a physical item or digital picture to the Block Dispensary Blockchain list item sell price in bitcoin but opt in to sell item for other eligible crypto currencies! Upload photo write description name item another upload music or podcast to blockchain option below main item upload! top of page with all the links to other pages as in all the other pages! Item is submitted and then sent to human hash verification on sector one of the Block Dispensary blockchain!