Create approval.js
Approval page is to be approved for a postion as a human hasher on the send item to sell block on the Block Dispensary Blockchain simple registration detials with submit button guideline and policies explained in short detail. example explanation of what to expect of the procedure of human hash function on the Block Dispensary Blockchain approve item to Sell blockchain the first of four sector blockchains on the Block Dispensary Blockchain Network. The top of the approve page being the same as all the other pages with link redirects to the websites other pages!