Create Blockchain.js
an example of the blockchain page for block dispensary with download option to download file for the block dispensary blockchain explanation of the block dispensary blockchain and the four sector blocks and how they are hashed to complete the main block confirmation! original top page links and BLOCK DISPENSARY at the bottom with a logo for block dispensary! where it states public chain data for the block dispensary the public chain for the blockchain could be displayed on this page or redircted to a different javascript page or redirected to a different website for block dispensary blockchain explorer!